// app/core/models/news-item.ts
// Для работы с данными лучше создавать интерфейсы, особенно если это данные получаемые от API,
// Если данные получем от API, то интерфейсы лучше всего держать в папке core/models
// Если нет, то в папке с компонентом в отдельном файле
// Модель через class создаем только в том случае если нам надо какие-то данные полученные от API привести к какой-нибудь конкретной модели используемой на фронте
export interface NewsItem {
  purchase: string;
  done: boolean;
  price: number;
}